Koin added, rebuilding 03.05.22

+++ app/src/main/java/com/denis/recipebookandroid/model/dao/entities/CookingProcessEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class CookingProcessEntity(
     @PrimaryKey
-    val processId: Int,
+    val processId: Long,
     val processNumber: Int,
     val description: String,
 )
